+++ frontend/src/App.tsx
@@ -4,8 +4,7 @@ import { ThemeProvider } from '@mui/material/styles';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
-import { Footer } from './components';
-import Navbar from './components/Navbar';
+import { Footer, Navbar } from './components';
 import { Inquiry, Home } from './pages';
 import BCTheme from './styles';
