+++ frontend/src/components/Navbar/Navbar.tsx
@@ -1,11 +1,11 @@
 import { AppBar, Box, Toolbar } from '@mui/material';
-import Logo from '../assets/logo.png';
+import Logo from '../../assets/logo.png';
 
 type NavbarProps = {
-  children: React.ReactNode;
+  children?: React.ReactNode;
 };
 
-const Navbar = (props: NavbarProps) => {
+export const Navbar = (props: NavbarProps) => {
   const { children = null } = props;
 
   return (
@@ -19,9 +19,13 @@ const Navbar = (props: NavbarProps) => {
         />
         {children}
       </Toolbar>
-      <Box sx={{ height: 5 }} bgcolor="#FCBA19" />
+      <Box role="presentation" sx={{ height: 5 }} bgcolor="#FCBA19" />
     </AppBar>
   );
+};
+
+Navbar.defaultProps = {
+  children: null,
 };
 
 export default Navbar;
