+++ frontend/src/components/index.ts
@@ -1,9 +1,10 @@
+export * from './Footer/Footer';
+export * from './Navbar/Navbar';
+export * from './buttons/Button';
 export * from './dialogs/Dialog';
-export * from './dialogs/LoginDialog';
 export * from './dialogs/ErrorDialog';
+export * from './dialogs/LoginDialog';
 export * from './loading/Loading';
 export * from './table/TableContainer';
 export * from './tablecells/DateCell';
 export * from './tablecells/StatusCell';
-export * from './buttons/Button';
-export * from './Footer/Footer';
